Move parse function to selector
